add custom snippets element
