--- docs/snippets/getting_started/quickstart.py
@@ -1,16 +1,0 @@
-# @@@SNIPSTART getting_started_index_snippet_start
-import dlt
-
-data = [
-    {'id': 1, 'name': 'Alice'},
-    {'id': 2, 'name': 'Bob'}
-]
-
-pipeline = dlt.pipeline(
-    pipeline_name='quick_start',
-    destination='duckdb',
-    dataset_name='mydata'
-)
-load_info = pipeline.run(data, table_name="users")
-print(load_info)
-# @@@SNIPEND
