+++ docs/website/docs/assets/invoices/invoice_2.txt
