--- docs/website/watch_snippets.js
@@ -1,11 +1,0 @@
-var watch = require('node-watch');
-const { exec } = require('child_process');
-
-console.log("Starting to watch for snippet changes")
-watch('../snippets', { recursive: true, filter: /\.py$/  }, function(evt, name) {
-  console.log('%s changed, updating snippets.', name);
-  exec('yarn snipsync', (err, stdout, stderr) => {
-    if (err) {
-      return;
-    }});
-});
